# Optimized fetch of random values using cache