@brendanashworth 2019-06-16
@TanukiSharp this is great, thank you! I've merged it in https://github.com/brendanashworth/generate-password/commit/ea642553c5ba327989f36d0b0f2d4e80ff25b45a and am releasing a patch release with the change.
Cheers 🍻